Fix code review findings.
