+++ extensions/sql/policy/docs/schema.sql
@@ -1,5 +1,5 @@
 --
---  Copyright (c) 2022 Daimler TSS GmbH
+--  Copyright (c) 2022 ZF Friedrichshafen AG
 --
 --  This program and the accompanying materials are made available under the
 --  terms of the Apache License, Version 2.0 which is available at
@@ -8,21 +8,32 @@
 --  SPDX-License-Identifier: Apache-2.0
 --
 --  Contributors:
---       Daimler TSS GmbH - Initial SQL Query
+--       ZF Friedrichshafen AG - Initial SQL Query
 --
+
+-- Statements are designed for and tested with Postgres only!
 
 -- table: edc_policies
 CREATE TABLE IF NOT EXISTS edc_policies
 (
-    policy_id VARCHAR(255) NOT NULL,
-    permissions VARCHAR(MAX),
-    prohibitions VARCHAR(MAX),
-    duties VARCHAR(MAX),
-    extensible_properties VARCHAR(MAX),
-    inherits_from VARCHAR(MAX),
-    assigner VARCHAR(MAX),
-    assignee VARCHAR(MAX),
-    target VARCHAR(MAX),
-    policy_type VARCHAR(MAX) NOT NULL,
+    policy_id VARCHAR NOT NULL,
+    permissions VARCHAR,
+    prohibitions VARCHAR,
+    duties VARCHAR,
+    extensible_properties VARCHAR,
+    inherits_from VARCHAR,
+    assigner VARCHAR,
+    assignee VARCHAR,
+    target VARCHAR,
+    policy_type VARCHAR NOT NULL,
     PRIMARY KEY (policy_id)
 );
+
+COMMENT ON COLUMN edc_policies.permissions IS 'Java List<Permission> serialized as JSON';
+COMMENT ON COLUMN edc_policies.prohibitions IS 'Java List<Prohibition> serialized as JSON';
+COMMENT ON COLUMN edc_policies.duties IS 'Java List<Duty> serialized as JSON';
+COMMENT ON COLUMN edc_policies.extensible_properties IS 'Java Map<String, Object> serialized as JSON';
+COMMENT ON COLUMN edc_policies.policy_type IS 'Java PolicyType serialized as JSON';
+
+CREATE UNIQUE INDEX edc_policies_id_uindex
+    ON edc_policies (id);
